+++ africanus/rime/bright.py
@@ -7,7 +7,7 @@ from __future__ import print_function
 
 import numpy as np
 
-from ..util import corr_shape as corr_shape_fn
+from ..util.shapes import corr_shape as corr_shape_fn
 
 
 def brightness(stokes, polarisation_type=None, corr_shape=None):
